fix(image): share responses text parsing
